TextDrawable and ProgressbarDrawable added. Some changes to graphics. Now using double instead of float in most of classes.

+++ src/main/java/utils/PointF.java
@@ -2,31 +2,31 @@ package utils;
 
 public class PointF
 {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public PointF(float x, float y)
+    public PointF(double x, double y)
     {
         this.x = x;
         this.y = y;
     }
 
-    public float getX()
+    public double getX()
     {
         return x;
     }
 
-    public void setX(float x)
+    public void setX(double x)
     {
         this.x = x;
     }
 
-    public float getY()
+    public double getY()
     {
         return y;
     }
 
-    public void setY(float y)
+    public void setY(double y)
     {
         this.y = y;
     }
